broken test/test_3d_fullscreen.py in smaller units to facilitate testing
The smaller units were:
	test_wavestiles3d.py -> existed before; simplified
	test_wavestiles3d_fullscreen.py -> stress the base unit with fullscreen toggles
	test_wavestiles3d_reversed.py -> stress the base unit with reverse oper

	test_shuffletiles.py -> existed before; simplified
	test_shuffletiles_fullscreen.py -> stress the base unit with fullscreen toggles
	test_shuffletiles_reversed.py -> stress the base unit with reverse oper
	
What the reduced cases don't show is if action + Reverse(action) have a smooth transition between actions.
But that was not appreciated in the original test because the scene action is too fast.

Using autotest snapshots for the forward and backward actions should be straightforward to asses smoothness.



git-svn-id: http://los-cocos.googlecode.com/svn/trunk@1183 f663ce52-ac46-0410-b8de-c1c220b0eb76